feat(location): integrate LocationManagerApiHooks for enhanced location management
- Added LocationManagerApiHooks to ModuleEntry for improved location handling.
- Removed deprecated hookLocationManager method from LocationApiHooks to streamline functionality.
