+++ app/src/main/java/com/noobexon/xposedfakelocation/xposed/hooks/LocationApiHooks.kt
@@ -1,6 +1,5 @@
 package com.noobexon.xposedfakelocation.xposed.hooks
 
-import android.location.Location
 import android.os.Build
 import android.util.Log
 import com.noobexon.xposedfakelocation.xposed.utils.LocationUtil
@@ -14,7 +13,6 @@ class LocationApiHooks(private val module: XposedInterface, private val classLoa
 
     fun initHooks() {
         hookLocation()
-        hookLocationManager()
         module.log(Log.INFO, tag, "Instantiated hooks successfully")
     }
 
@@ -148,28 +146,4 @@ class LocationApiHooks(private val module: XposedInterface, private val classLoa
         }
     }
 
-    private fun hookLocationManager() {
-        try {
-            val locationManagerClass = Class.forName("android.location.LocationManager", false, classLoader)
-            val method = locationManagerClass.getDeclaredMethod("getLastKnownLocation", String::class.java)
-
-            module.hook(method).intercept { chain ->
-                val original = chain.proceed() as? Location
-                module.log(Log.INFO, tag, "Leaving method getLastKnownLocation(provider)")
-                module.log(Log.INFO, tag, "\t Original location: $original")
-                val provider = chain.getArg(0) as String
-                module.log(Log.INFO, tag, "\t Requested data from: $provider")
-                if (PreferencesUtil.getIsPlaying() == true) {
-                    val fakeLocation = LocationUtil.createFakeLocation(provider = provider)
-                    module.log(Log.INFO, tag, "\t Modified location: $fakeLocation")
-                    fakeLocation
-                } else {
-                    original
-                }
-            }
-
-        } catch (e: Exception) {
-            module.log(Log.ERROR, tag, "Error hooking LocationManager - ${e.message}")
-        }
-    }
 }
